requirements.txt
# Core Framework
fastapi==0.109.0
pydantic==2.5.0
uvicorn==0.27.0

# LLM Infrastructure
langchain==0.1.4
langchain-community==0.0.19

# Document Processing & RAG
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2

# Storage
sqlalchemy==2.0.25

# Testing
pytest==7.4.4
httpx==0.26.0
